First cut, mostly untested.
