+++ tools/rostopic/src/rostopic/verbs/find.py
@@ -1,0 +1,64 @@
+# Software License Agreement (BSD License)
+#
+# Copyright (c) 2008, Willow Garage, Inc.
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#
+#  * Redistributions of source code must retain the above copyright
+#    notice, this list of conditions and the following disclaimer.
+#  * Redistributions in binary form must reproduce the above
+#    copyright notice, this list of conditions and the following
+#    disclaimer in the documentation and/or other materials provided
+#    with the distribution.
+#  * Neither the name of Willow Garage, Inc. nor the names of its
+#    contributors may be used to endorse or promote products derived
+#    from this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+# COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+
+from __future__ import division, print_function
+
+import sys
+
+
+def find_by_type(topic_type):
+    """
+    Lookup topics by topic_type
+    :param topic_type: type of topic to find, ``str``
+    :returns: list of topic names that use topic_type, ``[str]``
+    """
+    master = rosgraph.Master('/rostopic')
+    try:
+        t_list = _master_get_topic_types(master)
+    except socket.error:
+        raise ROSTopicIOException("Unable to communicate with master!")
+    return [t_name for t_name, t_type in t_list if t_type == topic_type]
+
+def _rostopic_cmd_find(argv=sys.argv):
+    """
+    Implements 'rostopic type'
+    :param argv: command-line args, ``[str]``
+    """
+    args = argv[2:]
+    from optparse import OptionParser
+    parser = OptionParser(usage="usage: %prog find msg-type", prog=NAME)
+    options, args = parser.parse_args(args)
+    if not len(args):
+        parser.error("please specify a message type")
+    if len(args) > 1:
+        parser.error("you may only specify one message type")
+    print('\n'.join(find_by_type(args[0])))
